fixed add album and song. fixed delete buttons. added admin only page

+++ src/main/java/com/kdg/springprojt5/controllers/api/ArtistApiController.java
@@ -2,8 +2,11 @@ package com.kdg.springprojt5.controllers.api;
 
 import com.kdg.springprojt5.controllers.api.dto.ArtistDto;
 import com.kdg.springprojt5.controllers.api.dto.NewArtistDto;
+import com.kdg.springprojt5.domain.Album;
 import com.kdg.springprojt5.domain.Artist;
+import com.kdg.springprojt5.service.AlbumService;
 import com.kdg.springprojt5.service.ArtistService;
+import com.kdg.springprojt5.service.SongService;
 import jakarta.validation.Valid;
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
@@ -22,10 +25,14 @@ public class ArtistApiController {
 
     private final Logger logger;
     private final ArtistService artistService;
+    private final AlbumService albumService;
+    private final SongService songService;
 
     private final ModelMapper modelMapper;
 
-    public ArtistApiController(ArtistService artistService, ModelMapper modelMapper) {
+    public ArtistApiController(ArtistService artistService, AlbumService albumService, SongService songService, ModelMapper modelMapper) {
+        this.albumService = albumService;
+        this.songService = songService;
         this.modelMapper = modelMapper;
         this.logger = LoggerFactory.getLogger(this.getClass().getName());
         this.artistService = artistService;
@@ -72,6 +79,15 @@ public class ArtistApiController {
 
     @DeleteMapping("/artist/{id}/delete")
     public ResponseEntity<Void> deleteArtist(@PathVariable long id) {
+        Artist artist = artistService.getArtistById(id);
+        List<Album> artistAlbums = artist.getAlbums();
+        if (artistAlbums == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        for (Album album : artistAlbums) {
+            songService.getSongsByAlbumId(album.getId()).forEach(song -> songService.deleteSong(song.getId()));
+            albumService.deleteAlbum(album.getId());
+        }
         artistService.deleteArtist(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
